+++ app/Http/Controllers/TahunAjaranController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\BAP;
 use App\Models\TahunAjaran;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
@@ -114,12 +115,17 @@ class TahunAjaranController extends Controller
     public function destroy(TahunAjaran $tahunAjaran)
     {
         $cek = TahunAjaran::where('unique', $tahunAjaran->unique)->first();
+        $cek_bap = BAP::where('tahun_ajaran_unique', $tahunAjaran->unique)->first();
 
         if ($cek->status == '1') {
             return response()->json(['errors' => 'Tahun Ajaran Aktif Tidak Bisa Dihapus!']);
         } else {
-            TahunAjaran::where('unique', $tahunAjaran->unique)->delete();
-            return response()->json(['success' => 'Data Tahun Ajaran Berhasil Dihapus!']);
+            if ($cek_bap) {
+                return response()->json(['errors' => 'Tahun Ajaran Tidak Bisa Dihapus!']);
+            } else {
+                TahunAjaran::where('unique', $tahunAjaran->unique)->delete();
+                return response()->json(['success' => 'Data Tahun Ajaran Berhasil Dihapus!']);
+            }
         }
     }
